requirements.txt
absl-py==1.0.0
art==5.4
backcall==0.2.0
blis==0.7.5
bpemb==0.3.3
cachetools==4.2.4
catalogue==2.0.6
catalyst==20.8.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
conllu==4.4.1
cycler==0.11.0
cymem==2.0.6
Cython==0.29.14
decorator==5.1.0
Deprecated==1.2.13
deprecation==2.1.0
filelock==3.3.2
flair==0.9
ftfy==6.0.3
future==0.18.2
gdown==3.12.2
gensim==3.8.3
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.41.1
huggingface-hub==0.1.2
hydra-slayer==0.3.0
hyperopt==0.2.5
idna==3.3
importlib-metadata==3.10.1
ipython==7.29.0
Janome==0.4.1
jedi==0.18.0
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
konoha==4.6.5
langcodes==3.3.0
langdetect==1.0.9
lxml==4.6.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
more-itertools==8.8.0
mpld3==0.3
murmurhash==1.0.6
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.1
overrides==3.1.0
packaging==21.2
pandas==1.1.5
parso==0.8.2
pathy==0.6.1
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.3.1
preshed==3.0.6
prompt-toolkit==3.0.22
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycm==3.3
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
segtok==1.5.10
sentencepiece==0.1.95
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlitedict==1.7.0
srsly==2.4.2
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
thinc==8.0.13
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchtext==0.11.0
torchvision==0.11.1+cu113
tqdm==4.62.3
traitlets==5.1.1
transformers==4.12.3
tweepy==4.3.0
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.9.0
wcwidth==0.2.5
Werkzeug==2.0.2
Wikipedia-API==0.5.4
wrapt==1.13.3
zipp==3.6.0
